requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
asyncpg==0.29.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
databases==0.9.0
fastapi==0.110.0
frozendict==2.4.0
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httptools==0.6.1
idna==3.6
lxml==5.1.0
multitasking==0.0.11
numpy==1.26.4
pandas==2.2.1
peewee==3.17.1
pipdeptree==2.16.1
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
yfinance==0.2.37
